+++ hehe-board/src/main/java/com/programmers/heheboard/domain/user/UserController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.programmers.heheboard.global.ApiResponse;
+import com.programmers.heheboard.global.response.ApiResponse;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
